updated posts and comments schemas

+++ server/src/data/migrations/20190630060319_posts.js
@@ -12,7 +12,7 @@ exports.up = function(knex, Promise) {
       .onUpdate("CASCADE");
 
     posts
-      .integer("created_at")
+      .bigInteger("created_at")
       .unsigned()
       .notNullable();
