+++ server/src/data/migrations/20190701105428_comments.js
@@ -21,7 +21,7 @@ exports.up = function(knex, Promise) {
       .onUpdate("CASCADE");
 
     comments
-      .integer("created_at")
+      .bigInteger("created_at")
       .unsigned()
       .notNullable();
